feat: Add --print flag for non-interactive output
This commit introduces a '--print' flag that allows the user to run the application in a non-interactive mode.

When the '--print' flag is provided, the application will:
- Fetch flight data from the OpenSky and FlightAware APIs.
- Print the flight data to standard output in JSON format.
- Print logs to standard output.
- Exit with code 0 on success.

If the '--print' flag is not provided, the application will start the HTTP server as usual.

This provides the flexibility to use the application as both a persistent service and a one-off command-line tool.